refactor(python-sdk): move authentication_header and extract_start_pid to envd/utils
Leaves e2b/envd/rpc.py purely about mapping RPC errors onto SDK
exceptions.
