+++ packages/python-sdk/e2b/sandbox_sync/filesystem/filesystem.py
@@ -23,11 +23,8 @@ from e2b.envd.api import (
     handle_envd_api_transport_exception_with_health,
 )
 from e2b.envd.filesystem import filesystem_connect, filesystem_pb
-from e2b.envd.rpc import (
-    authentication_header,
-    handle_rpc_exception_with_health,
-)
-from e2b.envd.utils import timeout_to_ms
+from e2b.envd.rpc import handle_rpc_exception_with_health
+from e2b.envd.utils import authentication_header, timeout_to_ms
 from e2b.envd.client_sync import create_rpc_client
 from e2b.envd.versions import (
     ENVD_DEFAULT_USER,
