+++ packages/python-sdk/e2b/sandbox_sync/filesystem/watch_handle.py
@@ -9,11 +9,8 @@ from e2b.envd.filesystem.filesystem_pb import (
     GetWatcherEventsRequest,
     RemoveWatcherRequest,
 )
-from e2b.envd.rpc import (
-    authentication_header,
-    handle_rpc_exception_with_health,
-)
-from e2b.envd.utils import timeout_to_ms
+from e2b.envd.rpc import handle_rpc_exception_with_health
+from e2b.envd.utils import authentication_header, timeout_to_ms
 from e2b.sandbox.filesystem.filesystem import map_entry_info
 from e2b.sandbox.filesystem.watch_handle import FilesystemEvent, map_event_type
